requirements.txt
pandas==1.4.2
numpy==1.21.5
matplotlib==3.5.1
matplotlib-venn==0.11.7
seaborn== 0.11.2
scipy==1.7.3
